#37 - Add state class
