+++ libs/noughts-and-crosses/src/lib/noughts-and-crosses.ts
@@ -1,4 +1,5 @@
-import { AgEngine, GameState } from "@lundin/age"
+import { AgEngine } from "@lundin/age"
+import { NoughtsAndCrossesState } from "./noughts-and-crosses-state"
 
 export class NoughtsAndCrosses {
 	private engine: AgEngine<NoughtsAndCrossesAction>
@@ -10,23 +11,5 @@ export class NoughtsAndCrosses {
 		this.engine.update(actions)
 	}
 }
-
-export class NoughtsAndCrossesState implements GameState {
-	constructor(
-		public board: NoughtsAndCrossesType[] = null,
-		public tick: number = 0,
-	) {
-		if (!board)
-			this.initialiseBoard()
-	}
-
-	private initialiseBoard() {
-		this.board = []
-		for (let i = 0; i < 9; i++)
-			this.board.push(NoughtsAndCrossesType.None)
-	}
-}
-
-export enum NoughtsAndCrossesType { None, Nought, Cross }
 
 export class NoughtsAndCrossesAction { }
